IssueCommand+PRCommand: Use emoji ID to correctly display custom emoji
This fixes the issue of discord not preprocessing bots' messages for
any custom emoji strings (ala :sadcaret:) and replacing them with the
ID format. This commit uses the ID format explicitly to fix that.

In the future it might be beneficial to use the discordjs api somehow
to not hardcode the emoji ID.