CommandParser: Remove mentions and emotes from commands
Sometimes it is useful to ping someone while running a command, i.e.

!pr <pr number> @user

However, currently the bot includes the mention in the search, causing
some annoyance.  This patch removes anything between <> that's a single
word, which both emotes and mentions are.

This could probably be refined further by looking into the mentions
object within the message, but this shouldn't cause any false positives
during normal usage.