Util+Commands: Cache emojis based on a globally configured guild
Previously the emoji cache wouldn't work in private message channels
due to it using a message's guild to resolve an emoji.

As Buggie is currently the only instance of the bot and as said
instance is only being used on the Serenity server there is no
problem in taking a guild as a config variable.

The emoji cache now also not only caches the emoji's name but the whole
reference object.

Resolves #21