Commands: Added the test262 command
This command fetches and outputs the results of the ecma262 tests
run in LibJS from https://libjs.dev/test262.

If no argument is given, the latest commit will be used. Otherwise
the first argument will be used to resolve a commit by hash.

If the first argument is "labels", an overview over the icons
used in the result output will be displayed.