QuickLinksCommand: Add the ability to suppress embeds for quick links
The link embeds take up significant visual space in discord, for a
preview of the github repro that we all know, there is no real reason
for it to be used for links, unless they are images / different sites.

The `!soytineres` command is special as it relies on the image emebed.