Commands: Migrate to discord native slash commands
This patch removes the old message based command handling code and
replaces it with the Interactions api.

Every implementation of the command superclass now has to provide
one or more command data objects which get registered as global commands
on startup. Global commands can take up to an hour to register in guilds
but they also work in direct messages.

Some commands have less aliases to reduce clutter in the command menu.

Error messages and the outputs of some commands only show up to user running
the command.