QuoteCommand: Fix test encoding of the /quote command (attempt #2)
This is paired back version of change I previously made in 37f927cc0c266d.
The changes to pass the encoding to the composeCreatePullRequest call,
end up making the request go to a BLOB API which has authentication
issues when running as the bot token.

After testing with the new unit test I added, I was able to validate
that all we need to do is change text encoding we use for the JSON
in order to make sure we don't corrupt the JSON file when we post
new records to it via the PR.

Fixes #356