test262Command: Get author name from commit.author if author is null
When a commit has an author with no associated GitHub account,
commit.author will be null. However, commit.commit.author will at least
give you the author name assigned to the commit. We can't get an avatar
URL or profile URL from this though, so the command will just display
the author name in this case.