Index: Remove blob and tree embeds from github.com
Repository embeds from discord fill up a lot of screen space without
providing much relevant information.

This patch adds a message listener that removes the embeds on a message
automatically if the message has an embed from github.com with blob or
tree being the third segment in the pathname (the type of github url).
Sadly the api does not allow more fine grain removal of a single embed,
however multiple embeds in a single message are quite rare.

As requested by linusg on discord:
https://discord.com/channels/830522505605283862/830739873119207426/986007342443282493