Move to the pnpm package manager
Yarn 1, what the project used before, has been deprecated for a while.
Updating to Yarn Berry (4) would be possible, but it is not that
widespread yet.
pnpm is another alternative package manager which many projects have now
switched to. It is a lot faster than yarn, and offers safer dependency
management